Funciones WiMOD LoRaWAN (para envio) separadas del main
Implementaciones del main_pic llevadas a WMLW_API
Cabeceras LoRaWAN_API y hci_stack modificadas en los prototipos,
implicando el cambio de las implementaciones respectivas

+++ source/WiMOD_LoRaWAN_API.h
@@ -61,11 +61,11 @@ extern "C" {
 
     // send unconfirmed radio data
     int
-    WiMOD_LoRaWAN_SendURadioData(UINT8 port, UINT8 *data, int length);
+    WiMOD_LoRaWAN_SendURadioData(UINT8 port, UINT8 *data, UINT8 length);
 
     // send confirmed radio data
     int
-    WiMOD_LoRaWAN_SendCRadioData(UINT8 port, UINT8 *data, int length);
+    WiMOD_LoRaWAN_SendCRadioData(UINT8 port, UINT8 *data, UINT8 length);
 
     // receiver process
     void
